Properly load native library
The native xrootd library is stored in a .nar file, which can't be
directly loaded via System.loadLibrary() .. fortunately the NAR plugin
provides a helper method to load shared libraries from nar archives in
an architechture-independent way. Use that.

This has the unfortunate effect of being incompatible with the
makefile-based build system since the NarSystem library loader won't
exist.